+++ tailwind.config.js
@@ -37,11 +37,11 @@ module.exports = {
           "0%": { transform: "translate3d(0,-10%,0)", opacity: 0 },
           "100%": { transform: "translateZ(0)", opacity: 1 },
         },
-        slideDown: {
+        accordionSlideDown: {
           from: { height: 0 },
           to: { height: "var(--radix-accordion-content-height)" },
         },
-        slideUp: {
+        accordionSlideUp: {
           from: { height: "var(--radix-accordion-content-height)" },
           to: { height: 0 },
         },
@@ -54,8 +54,10 @@ module.exports = {
         hideLeft: "hideLeft 0.3s cubic-bezier(0.19, 1, 0.22, 1)",
         fadeInUp: "fadeInUp 0.2s ease-in-out",
         fadeInDown: "fadeInDown 0.2s ease-in-out",
-        slideDown: "slideDown 300ms cubic-bezier(0.87, 0, 0.13, 1)",
-        slideUp: "slideUp 300ms cubic-bezier(0.87, 0, 0.13, 1)",
+        accordionSlideDown:
+          "accordionSlideDown 300ms cubic-bezier(0.87, 0, 0.13, 1)",
+        accordionSlideUp:
+          "accordionSlideUp 300ms cubic-bezier(0.87, 0, 0.13, 1)",
       },
     },
   },